backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
mysqlclient==2.2.8
requests==2.33.1
python-dotenv==1.2.2
django-autoslug==1.9.9